refactor: Rename WeakForm to functional
