+++ src/serac/physics/utilities/functional/polynomials.hpp
