+++ src/serac/physics/utilities/functional/tests/CMakeLists.txt
@@ -5,18 +5,18 @@
 # SPDX-License-Identifier: (BSD-3-Clause)
 
 # Then add the examples/tests
-set(weakform_tests_serial
+set(functional_tests_serial
     hcurl_unit_tests.cpp
     test_tensor_ad.cpp
     tuple_arithmetic_unit_tests.cpp)
     
-foreach(filename ${weakform_tests_serial})
+foreach(filename ${functional_tests_serial})
     get_filename_component(test_name ${filename} NAME_WE)
     
     blt_add_executable(NAME        ${test_name}
                        SOURCES     ${filename}
                        OUTPUT_DIR  ${TEST_OUTPUT_DIRECTORY}
-                       DEPENDS_ON  gtest serac_weakform ${weakform_depends}
+                       DEPENDS_ON  gtest serac_functional ${functional_depends}
                        FOLDER      serac/tests )
     blt_add_test(NAME          ${test_name}
                  COMMAND       ${test_name}
@@ -24,18 +24,18 @@ foreach(filename ${weakform_tests_serial})
 endforeach()
 
 # Then add the examples/tests
-set(weakform_tests_mpi
-    # weak_form_boundary_test.cpp
-    weak_form_comparisons.cpp
+set(functional_tests_mpi
+    # functional_boundary_test.cpp
+    functional_comparisons.cpp
     )
     
-foreach(filename ${weakform_tests_mpi})
+foreach(filename ${functional_tests_mpi})
     get_filename_component(test_name ${filename} NAME_WE)
     
     blt_add_executable(NAME        ${test_name}
                        SOURCES     ${filename}
                        OUTPUT_DIR  ${TEST_OUTPUT_DIRECTORY}
-                       DEPENDS_ON  gtest serac_weakform ${weakform_depends}
+                       DEPENDS_ON  gtest serac_functional ${functional_depends}
                        FOLDER      serac/tests )
     blt_add_test(NAME          ${test_name}
                  COMMAND       ${test_name}
@@ -44,31 +44,31 @@ endforeach()
 
 
 # not sure how we want to handle performance-related tests
-set(weakform_performance_tests
+set(functional_performance_tests
     performance/J2_material_with_AD.cpp
     performance/diffusion_performance_test.cpp)
  
-foreach(filename ${weakform_performance_tests})
+foreach(filename ${functional_performance_tests})
     get_filename_component(test_name ${filename} NAME_WE)
     blt_add_executable(NAME        ${test_name}
                        SOURCES     ${filename}
                        OUTPUT_DIR  ${TEST_OUTPUT_DIRECTORY}
-                       DEPENDS_ON  gtest serac_weakform ${weakform_depends}
+                       DEPENDS_ON  gtest serac_functional ${functional_depends}
                        FOLDER      serac/tests )
 endforeach()
 
 if(ENABLE_CUDA)
 
-    set(weakform_tests_cuda
+    set(functional_tests_cuda
         tensor_unit_tests_cuda.cu)
         
-    foreach(filename ${weakform_tests_cuda})
+    foreach(filename ${functional_tests_cuda})
         get_filename_component(test_name ${filename} NAME_WE)
         
         blt_add_executable(NAME        ${test_name}
                            SOURCES     ${filename}
                            OUTPUT_DIR  ${TEST_OUTPUT_DIRECTORY}
-                           DEPENDS_ON  gtest serac_weakform ${weakform_depends} cuda
+                           DEPENDS_ON  gtest serac_functional ${functional_depends} cuda
                            FOLDER      serac/tests)
         blt_add_test(NAME          ${test_name}
                      COMMAND       ${test_name}
