+++ src/serac/physics/utilities/functional/tests/tensor_unit_tests.cpp
@@ -4,8 +4,8 @@
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
-#include "serac/physics/utilities/weak_form/detail/metaprogramming.hpp"
-#include "serac/physics/utilities/weak_form/tensor.hpp"
+#include "serac/physics/utilities/functional/detail/metaprogramming.hpp"
+#include "serac/physics/utilities/functional/tensor.hpp"
 
 using namespace serac;
